requirements.txt
# ── Web Framework ─────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.1.0

# ── Config ────────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# ── LangChain / LangGraph ─────────────────────────────────────────
langchain>=0.3.0
langchain-core>=1.0.0
langchain-openai>=1.1.10
langchain-groq>=0.2.0
langchain-community>=0.4.1
langchain-tavily>=0.2.0
langgraph>=1.0.10
langgraph-checkpoint-postgres>=3.0.4

# ── LLM Providers ─────────────────────────────────────────────────
openai>=1.51.0
groq>=0.11.0

# ── HTTP / async ──────────────────────────────────────────────────
httpx>=0.27.0
tenacity>=8.2.0

# ── CrewAI (Stock Analysis) ───────────────────────────────────────
crewai>=0.80.0

# ── Stock Data ────────────────────────────────────────────────────
yfinance>=0.2.40
numpy>=1.26.0
pandas>=2.1.0

# ── Vector / RAG (Chatbot) ────────────────────────────────────────
faiss-cpu>=1.8.0
tiktoken>=0.7.0
pypdf>=4.3.1
beautifulsoup4>=4.12.3
python-docx>=1.1.0
openpyxl>=3.1.0
pdf2image>=1.17.0
Pillow>=10.0.0

# ── Database ──────────────────────────────────────────────────────
SQLAlchemy>=2.0.36
aiosqlite>=0.20.0
greenlet>=3.0.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
redis>=5.0.8
pymongo>=4.6.0
motor>=3.3.0
PyGithub>=2.8.1

# ── Observability ─────────────────────────────────────────────────
langsmith>=0.2.0

# ── PDF / Document ────────────────────────────────────────────────
PyPDF2>=3.0.0
llmwhisperer-client>=2.3.0

# ── Browser Automation ────────────────────────────────────────────
browser-use>=0.1.40

# ── Testing ───────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
pytest-cov>=5.0.0
